pyproject.toml
[project]
name = "smart-routing-platform"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.129.0",
    "openpyxl>=3.1.5",
    "ortools>=9.15.6755",
    "pandas-stubs~=3.0.0",
    "psycopg2>=2.9.11",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "sqlalchemy[asyncio]>=2.0.46",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.4",
]

[tool.ruff]
line-length = 100
target-version = "py314"
fix = true
show-fixes = true

exclude = [
    ".venv",
    "venv",
    "__pycache__",
    "alembic/versions",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle
    "F",      # pyflakes (импорты, undefined)
    "I",      # сортировка импортов
    "B",      # bugbear (реальные баги)
    "UP",     # pyupgrade
    "ASYNC",  # async ошибки
    "SIM",    # упрощения
]

ignore = [
    "E501",   # длина строки (мы уже контролируем 100)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
